+++ database/migrations/2019_01_23_021513_create_campaigns_table.php
@@ -14,9 +14,9 @@ class CreateCampaignsTable extends Migration
     public function up()
     {
         Schema::create('campaigns', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('uuid');
-            $table->integer('squad_id')->unsigned();
+            $table->bigInteger('squad_id')->unsigned();
             $table->integer('continent_id')->unsigned();
             $table->integer('week_id')->unsigned();
             $table->timestamps();
